Sortir les données du seed élections dans un module pur
Le test importait le script, donc le singleton Prisma, qui échoue sans
DATABASE_URL : vert en local grâce au dotenv du script, rouge en CI.
